+++ tests/Feature/WhereConditionsTest.php
@@ -34,7 +34,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "id"
                             comparator: IN
                             value: ["1", "3", "5", "7", "9"]
@@ -101,7 +101,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "id"
                             comparator: NOT_IN
                             value: ["1", "3", "5", "7", "9"]
@@ -168,7 +168,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "id"
                             comparator: EQUAL
                             value: "2"
@@ -207,7 +207,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "id"
                             comparator: NOT_EQUAL
                             value: "2"
@@ -302,7 +302,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "price"
                             comparator: LESS_THAN
                             value: "15"
@@ -355,7 +355,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "price"
                             comparator: LESS_THAN_EQUAL
                             value: "15"
@@ -415,7 +415,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "price"
                             comparator: MORE_THAN
                             value: "15"
@@ -489,7 +489,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "price"
                             comparator: MORE_THAN_EQUAL
                             value: "15"
@@ -570,7 +570,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: {
+                        where: {
                             column: "description"
                             comparator: LIKE
                             value: "%lo%"
@@ -623,6 +623,135 @@ class WhereConditionsTest extends TestCase
         ], $result);
     }
 
+    public function testWhereNotLikeCondition(): void
+    {
+        $result = $this->graphQL->executeQuery(
+            <<<GQL
+            {
+                products (
+                    conditions: {
+                        where: {
+                            column: "description"
+                            comparator: NOT_LIKE
+                            value: "%lo%"
+                        }
+                    }
+                ) {
+                    id
+                    category_id
+                    name
+                    description
+                    price
+                }
+            }
+            GQL
+        );
+
+        $this->assertSame([
+            'data' => [
+                'products' => [
+                    [
+                        'id' => 5,
+                        'category_id' => 3,
+                        'name' => 'Product #5',
+                        'description' => 'Aster mor',
+                        'price' => 16.2,
+                    ],
+                    [
+                        'id' => 10,
+                        'category_id' => 3,
+                        'name' => 'Product #10',
+                        'description' => 'Mora de sito',
+                        'price' => 22.0,
+                    ],
+                ],
+            ],
+        ], $result);
+    }
+
+    public function testWhereOnHasManyCondition(): void
+    {
+        $result = $this->graphQL->executeQuery(
+            <<<GQL
+            {
+                categories {
+                    id
+                    name
+                    products__category_id (
+                        conditions: {
+                            where: {
+                                column: "description"
+                                value: null
+                                comparator: NOT_EQUAL
+                            }
+                        }
+                    ) {
+                        id
+                        name
+                        description
+                        price
+                    }
+                }
+            }
+            GQL
+        );
+
+        $this->assertSame([
+            'data' => [
+                'categories' => [
+                    [
+                        'id' => '1',
+                        'name' => 'Category #1',
+                        'products__category_id' => [
+                            [
+                                'id' => 1,
+                                'name' => 'Product #1',
+                                'description' => 'Lorem ipsum',
+                                'price' => 6.99,
+                            ],
+                            [
+                                'id' => 2,
+                                'name' => 'Product #2',
+                                'description' => 'Dolor sit',
+                                'price' => 5.69,
+                            ],
+                        ],
+                    ],
+                    [
+                        'id' => '2',
+                        'name' => 'Category #2',
+                        'products__category_id' => [
+                            [
+                                'id' => 4,
+                                'name' => 'Product #4',
+                                'description' => 'Dolor sit',
+                                'price' => 12.49,
+                            ],
+                        ],
+                    ],
+                    [
+                        'id' => '3',
+                        'name' => 'Category #3',
+                        'products__category_id' => [
+                            [
+                                'id' => 5,
+                                'name' => 'Product #5',
+                                'description' => 'Aster mor',
+                                'price' => 16.2,
+                            ],
+                            [
+                                'id' => 10,
+                                'name' => 'Product #10',
+                                'description' => 'Mora de sito',
+                                'price' => 22.0,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ], $result);
+    }
+
     public function testMultipleAndWhereConditions(): void
     {
         $result = $this->graphQL->executeQuery(
@@ -630,7 +759,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: [
+                        where: [
                             {
                                 column: "id"
                                 comparator: IN
@@ -676,7 +805,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        or_where: [
+                        where_or: [
                             {
                                 column: "id"
                                 comparator: IN
@@ -736,9 +865,71 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: [
+                        where_or: [
                             {
-                                or_where: [
+                                where: [
+                                    {
+                                        column: "id"
+                                        comparator: IN
+                                        value: ["1", "3"]
+                                    }
+                                    {
+                                        column: "id"
+                                        comparator: IN
+                                        value: ["3", "5"]
+                                    }
+                                ]
+                            }
+                            {
+                                column: "id"
+                                comparator: IN
+                                value: ["7"]
+                            }
+                        ]
+                    }
+                ) {
+                    id
+                    category_id
+                    name
+                    description
+                    price
+                }
+            }
+            GQL
+        );
+
+        $this->assertSame([
+            'data' => [
+                'products' => [
+                    [
+                        'id' => 3,
+                        'category_id' => null,
+                        'name' => 'Product #3',
+                        'description' => null,
+                        'price' => 18.99,
+                    ],
+                    [
+                        'id' => 7,
+                        'category_id' => null,
+                        'name' => 'Product #7',
+                        'description' => 'Loras dore',
+                        'price' => 15.0,
+                    ],
+                ],
+            ],
+        ], $result);
+    }
+
+    public function testNestedWhereOrConditions(): void
+    {
+        $result = $this->graphQL->executeQuery(
+            <<<GQL
+            {
+                products (
+                    conditions: {
+                        where: [
+                            {
+                                where_or: [
                                     {
                                         column: "id"
                                         comparator: IN
@@ -791,7 +982,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: [
+                        where: [
                             {
                                 column: "id"
                                 comparator: NOT_IN
@@ -895,7 +1086,7 @@ class WhereConditionsTest extends TestCase
             {
                 products (
                     conditions: {
-                        and_where: [
+                        where: [
                             {
                                 column: "id"
                                 comparator: IN